+++ CHANGELOG.md
@@ -53,6 +53,7 @@ More env variable names you can find in `ConnectionOptionsEnvReader` class.
 of `QueryRunner`, e.g. `queryRunner.connection` and `queryRunner.manager`
 * now `update` method in `QueryBuilder` accepts `Partial<Entity>` and property names used in update map are 
 column property names and they are automatically mapped to column names 
+* `SpecificRepository` has been removed. Instead new `RelationQueryBuilder` was introduced.
 
 ### DEPRECATIONS